add internal class for manager wire API
